+++ extension/src/setup/autoInstall.ts
@@ -1,9 +1,7 @@
-import { join, relative } from 'path'
 import { getPythonExecutionDetails } from '../extensions/python'
 import { findPythonBin, getDefaultPython, installPackages } from '../python'
 import { ConfigKey, getConfigValue } from '../vscode/config'
 import { getFirstWorkspaceFolder } from '../vscode/workspaceFolders'
-import { isSameOrChild } from '../fileSystem'
 import { Toast } from '../vscode/toast'
 
 export const findPythonBinForInstall = async (): Promise<
@@ -16,23 +14,6 @@ export const findPythonBinForInstall = async (): Promise<
   return findPythonBin(
     manualPython || autoPythonDetails?.join('') || getDefaultPython()
   )
-}
-
-export const getPythonBinDisplayText = (
-  path: string | undefined
-): string | undefined => {
-  if (!path) {
-    return
-  }
-
-  const workspaceRoot = getFirstWorkspaceFolder()
-  if (!workspaceRoot) {
-    return path
-  }
-
-  return isSameOrChild(workspaceRoot, path)
-    ? join('.', relative(workspaceRoot, path))
-    : path
 }
 
 const showInstallProgress = (
